requirements.txt
ultralytics==8.2.77
opencv-python>=4.8.0
PyYAML>=6.0
PySide6>=6.6.0

